% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tile_weight.R
\name{geom_tile_weight}
\alias{geom_tile_weight}
\title{geom_tile_complete: Draw a tile heatmap with automatic aggregation and log-transform}
\usage{
geom_tile_weight(mapping = NULL, data = NULL, position = "identity", ...)
}
\arguments{
\item{mapping, }{data, stat, position Standard ggplot2 arguments}

\item{...}{Additional arguments passed to layer()}
}
\description{
geom_tile_complete: Draw a tile heatmap with automatic aggregation and log-transform
}
