% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idid_colors_table}
\alias{idid_colors_table}
\title{A table with the main colors used in base plots in this package$}
\format{
A data frame with 644 rows and 3 variables:
\describe{
\item{pal}{palette name}
\item{base}{base color, for one color graphs}
\item{text}{color for text and axes}
\item{vector}{a vector of colors used in the scale}
...
}
}
\usage{
idid_colors_table
}
\description{
A table containing the name of the color combination, a base color,
a background color, a lighter and a darker one
}
\keyword{datasets}
