% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_contrib_stats.R
\name{idid_contrib_stats}
\alias{idid_contrib_stats}
\title{Compute quick descriptive statistics on the effective sample}
\usage{
idid_contrib_stats(
  reg,
  var_interest,
  contrib_threshold,
  threshold_change = 0.05,
  ...
)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{contrib_threshold}{A numeric (optional). Weight below which
observations are deemed to be non-contributing.
If not provided, will be determined by running
\code{ididvar::idid_contrib_threshold}}

\item{threshold_change}{A double (between 0 and 1). The change threshold in
estimate and s.e. when dropping observations.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A list with 4 entries:
\itemize{
\item \code{n_total}: the total sample size, before dropping observations with
missing values
\item \code{n_nominal}: the nominal sample size, ie the number of observations in the
regression
\item \code{n_effective}: the effective sample size, after removing non contributing
observations
\item \code{prop_effective}: the ratio of the effective to the nominal sample sizes
}
}
\description{
Reports the total, nominal and effective sample size, as well as the
proportion of observations from the nominal sample that are part of the
effective one.
}
\examples{
reg_ex <- ggplot2::txhousing |>
  lm(formula = volume ~ sales + listings + city + as.factor(date))

idid_contrib_stats(reg_ex, "sales")
}
