% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_contrib_threshold.R
\name{idid_contrib_threshold}
\alias{idid_contrib_threshold}
\title{Find a weight threshold below which observations do not contribute}
\usage{
idid_contrib_threshold(
  reg,
  var_interest,
  threshold_change = 0.05,
  search_step = 0.05
)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The treatment or main variable of interest.}

\item{threshold_change}{A number (between 0 and 1).
The maximum size of change in predictions allowed after dropping participants (i.e., weighting participants 0). Optional. Default is \code{0.05}, which means up to 5\% of observations can be dropped without exceeding the threshold change.}

\item{search_step}{A fraction of observations or weight values. Optional. Default is \code{0.05}.}
}
\value{
The last value from the weights at the point where no more participants could be "dropped" without significantly altering the estimates. It's returned invisibly, i.e., it's only visible if you call this function with \code{invisible()} around \code{idid_contrib_threshold} in your code.
}
\description{
Find a weight threshold below which removing observations does not change
the point estimate or the standard error of the estimate of interest by more
than a given proportion.
}
