% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_contrib_viz.R
\name{idid_contrib_viz}
\alias{idid_contrib_viz}
\title{Compute heatmap contribution weights}
\usage{
idid_contrib_viz(
  reg,
  var_interest,
  var_1,
  var_2,
  contrib_threshold,
  keep_labels = TRUE,
  ...
)
}
\arguments{
\item{reg}{A regression object created from \code{ididvar::create_regression()} function.}

\item{var_interest}{The variable that we are identifying using the idid method. This is typically one of the dependent variables in the regression.}

\item{var_1}{The first variable shown on the x-axis. This argument represents a set of characteristics for "all" observations in an area.}

\item{var_2}{(Optional) A second binary or continuous variable to be shown on the y-axis, with values indicating the number of times each value appears within \code{var_interest}.}

\item{keep_labels}{(Optional) If this is set to TRUE, we will include labels for all relevant areas; otherwise we only display a plot.}

\item{threshold_diff}{(Optional) A numeric value between 0 and 1 representing the maximum proportion of data points that can be removed before we suspect that our estimate is changing more than by a \code{threshold_diff} percentage.}
}
\value{
A ggplot object representing contribution heatmap. The function will print a warning message if \code{threshold_diff} is too low and not enough data would remain by dropping non-contributing cases.
}
\description{
Create a heatmap showing which observations contribute to the identification of certain variables.
}
