% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_diff_drop.R
\name{idid_diff_drop}
\alias{idid_diff_drop}
\title{Difference drop regression output}
\usage{
idid_diff_drop(reg, var_interest, prop_keep, threshold_diff = 0.05)
}
\arguments{
\item{reg}{A regression result object.}

\item{var_interest}{The variable to examine on \code{reg}.}

\item{prop_keep}{Proportion (as a proportion, e.g., 0.2 for 20\%) of "largest" values to keep.}

\item{threshold_diff}{What constitutes an implausible difference or uncertainty? Optional.

If the differences or standard errors are large, drop their corresponding observation from the dataset.
The function uses these in \code{ididvar::idid_weights}.}
}
\value{
A dataframe of interest:
\itemize{
\item \code{prop_keep}: the proportion (of total data) kept by calling this method.
\item \code{var_est}: relative change between \code{est} with and without large values
\item \code{var_see}: relative change between their standard errors
}
}
\description{
Compute a difference in coefficients and their standard errors for an item of interest.
If these are not "small," i.e. below the threshold, drop it.
}
