% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_drop_change.R
\name{idid_drop_change}
\alias{idid_drop_change}
\title{Variation in estimate when dropping low weight observations}
\usage{
idid_drop_change(reg, var_interest, prop_drop)
}
\arguments{
\item{reg}{A regression result object.}

\item{var_interest}{The variable to examine on \code{reg}.}

\item{prop_drop}{Proportion of observations to drop

If the differences or standard errors are large, drop their corresponding observation from the dataset.
The function uses these in \code{ididvar::idid_weights}.}
}
\value{
A dataframe of interest:
\itemize{
\item \code{prop_drop}: the proportion (of total data) dropped by calling this method.
\item \code{var_est}: relative change between \code{est} with and without large values
\item \code{var_se}: relative change between their standard errors
}
}
\description{
Computes a variation of the point estimate and standard errors of the variable
of interest when dropping low weight observations.
}
