% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_drop_change_viz.R
\name{idid_drop_change_viz}
\alias{idid_drop_change_viz}
\title{Drop change}
\usage{
idid_drop_change_viz(
  reg,
  var_interest,
  search_step = 0.05,
  search_start = search_step,
  search_end = 1 - search_step,
  threshold_change = 0.05
)
}
\arguments{
\item{reg}{A linear model. Optional.}

\item{var_interest}{A single variable. Optionally, a vector of variables if \code{reg_is_poly} is TRUE in the linear model used for prediction.}

\item{search_step}{One of \code{0}, \code{'0'}, \code{Inf} or \code{.Machine$double.xmax}. Currently unused; must be specified if it isn't one of these values or an error will occur. Optional.}

\item{search_start}{A positive number between 0 and .99 (inclusive). Optional.}

\item{search_end}{A positive number between 1 and .9999 (inclusive) and different from \code{search_start}. Currently unused; must not be equal to \code{.Machine$double.xmax} or an error will occur. Optional.}
}
\value{
If successful, a plot is made with lines representing the proportion of change in the measure(s) of interest at various drop thresholds as compared to the full sample from one end to the other. Invisible otherwise.
}
\description{
A short description...
}
