% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_drop_reg.R
\name{idid_drop_reg}
\alias{idid_drop_reg}
\title{Run the regression with low weight observations dropped}
\usage{
idid_drop_reg(reg, var_interest, prop_drop, ...)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{prop_drop}{Proportion of observations to drop}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A regression object
}
\description{
Re-run the regression but dropping low weight observations.
}
\details{
This function recomputes the weights
}
\examples{
reg_ex <- ggplot2::economics |>
  lm(formula = unemploy ~ pce + uempmed + psavert + pop)

idid_drop_reg(reg_ex, "pce", prop_drop = 0.1)
}
