% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_var_group.R
\name{idid_grouping_var}
\alias{idid_grouping_var}
\title{Find the variable along which the weights are the most homogeneous
within groups formed by this variable}
\usage{
idid_grouping_var(reg, var_interest, grouping_vars = "everything")
}
\arguments{
\item{reg}{An object describing the model (e.g. \code{lm()} or \code{glm()}) fit.}

\item{var_interest}{The variable of interest.}

\item{grouping_vars}{A list or vector of variables to use as the grouping factor.}
}
\value{
Data frame where each row is a unique group and contains its mean variance,
ordered from lowest to highest. Currently unused arguments are passed on to
\code{idid_weights}.
}
\description{
For now, I
}
