% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_to_cf.R
\name{idid_to_cf}
\alias{idid_to_cf}
\alias{idid_to_cf.fixest}
\alias{idid_to_cf.ivreg}
\alias{idid_to_cf.default}
\title{Transform an IV regression into its control function equivalent}
\usage{
idid_to_cf(reg, ...)

\method{idid_to_cf}{fixest}(reg, ...)

\method{idid_to_cf}{ivreg}(reg, ...)

\method{idid_to_cf}{default}(reg, ...)
}
\arguments{
\item{reg}{An IV regression object. Currently supported classes are
\code{fixest::fixest} and \code{AER::ivreg}.}

\item{...}{Additional arguments passed to \code{stats::update} when refitting
the model.}
}
\value{
A model object of the same class as \code{reg} (\code{fixest} or \code{ivreg}),
representing the equivalent control-function specification. The returned
model uses the same data, controls, and variance specification as the
original regression.
}
\description{
Takes an instrumental variables (IV) regression estimated
with either \code{fixest::feols} or \code{AER::ivreg} and refits it
as a \emph{control function} (CF) specification.
}
\details{
The resulting model reproduces the second-stage regression where the
residuals from each first-stage (for endogenous regressors) are added as
additional controls. These residuals are appended to the data as
\verb{<var>_first} and added as regressors in the updated model.

This function allows to explore the identifying variation of
IV estimates using the other functions of the \code{ididvar} package.

The function ensures that covariance options (e.g., clustering, weights)
and other model arguments are preserved from the original regression where
possible. If none are provided, the standard errors will be incorrect.
}
\section{Methods (by class)}{
\itemize{
\item \code{idid_to_cf(fixest)}: Method for \code{fixest} (feols) objects

\item \code{idid_to_cf(ivreg)}: Method for \code{ivreg} objects

\item \code{idid_to_cf(default)}: Default method (unsupported model types)

}}
\examples{
\dontrun{
library(fixest)
library(AER)

# feols example
data(trade)
reg_feols <- feols(Euros ~ log(dist) | log(dist) ~ log(air) + log(dist), data = trade)
reg_feols_cf <- idid_to_cf(reg_feols)

# ivreg example
data("CigarettesSW")
reg_ivreg <- ivreg(log(packs) ~ log(rprice) + log(rincome) | log(rtax) + log(rincome),
                data = CigarettesSW)
reg_ivreg_cf <- idid_to_cf(reg_ivreg)
}

}
\seealso{
\link[fixest:feols]{fixest::feols}, \link[AER:ivreg]{AER::ivreg}, \link[stats:update]{stats::update}
}
\concept{idid_methods}
