% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_contrib.R
\name{idid_viz_contrib}
\alias{idid_viz_contrib}
\title{Visualization of observations contributing (or not) to estimation}
\usage{
idid_viz_contrib(
  reg,
  var_interest,
  var_x,
  var_y,
  contrib_threshold,
  threshold_change = 0.05,
  order = "",
  colors = c("#C25807", "#FBE2C5", "#300D49"),
  keep_labels = TRUE,
  ...
)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{var_x}{A variable in the data set used in \code{reg} to plot on the
x-axis.}

\item{var_y}{A variable in the data set used in \code{reg} to plot on the
y-axis (optional). If not specified, produces a bar chart.}

\item{contrib_threshold}{A numeric (optional). Weight below which
observations are deemed to be non-contributing.
If not provided, will be determined by running
\code{ididvar::idid_contrib_threshold}}

\item{threshold_change}{A double (between 0 and 1). The change threshold in
estimate and s.e. when dropping observations.}

\item{order}{A string (either "x", "y" or "xy") describing whether the
graph should be order, along the x or y axis or both.
If anything else is specified, no specific ordering will be applied.}

\item{colors}{A string vector of colors for the palette.
I recommend to pass a vector of 3 distinct colors, with a lighter color in
the middle, constituting a diverging scale. It allows a clear distinction
between contributing and non contributing observations.}

\item{keep_labels}{A boolean (optional). If FALSE, removes y labels and ticks.
This option is useful for panels with a large number of individuals.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A ggplot2 graph of observations contributing to estimation.

If var_y is specified, returns a heatmap whose color describes whether a
given observation does not contribute (i.e. can be dropped, along with all
lower weight observations, without substantially altering the estimate)

If var_y is not specified, returns a bar chart representing the number of
observations that can be dropped in each group.
}
\description{
Make a graph to visualize observations that can be dropped without
changing the point estimate or the standard error of the estimate of
interest by more than a given proportion (\code{threshold_change}).

Makes a heatmap or a bar chart, depending on the number of dimensions
specified.
}
\examples{
reg_ex <- ggplot2::txhousing |>
  lm(formula = log(sales) ~ median + listings + city + as.factor(date))

idid_viz_contrib(reg_ex, "median", year, city, order = "y") +
  ggplot2::labs(x = NULL, y = NULL)

idid_viz_contrib(reg_ex, "median", var_y = city, order = "y") +
  ggplot2::labs(x = NULL)
}
