% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_weights.R
\name{idid_viz_weights}
\alias{idid_viz_weights}
\title{Visualization of Identifying Variation Weights}
\usage{
idid_viz_weights(reg, var_interest, var_x, var_y, keep_labels = TRUE)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{var_x}{A variable in the data set used in \code{reg} to plot on the
x-axis.}

\item{var_y}{A variable in the data set used in \code{reg} to plot on the
y-axis (optional). If not specified, produces a bar chart.}

\item{keep_labels}{A boolean (optional). If FALSE, remove y labels and ticks.}
}
\value{
A ggplot2 graph of identifying variation weights.

If var_y is specified, returns a heatmap whose color represents a
categorized version of the identifying variation weights (the categorization
prevents the vizualisation from being driven by extremely large weights).

If var_y is not specified, returns a bar chart representing the weights in
each group.
}
\description{
Makes a graph to visualize the identifying variation weights (a heatmap or
a bar chart, depending on the number of dimensions specified)
}
\examples{
reg_test <- ggplot2::economics |>
 lm(formula = unemploy ~ pce + uempmed + psavert + pop)

idid_weights(reg_test, "pce") |>
 head()
}
