% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_weights.R
\name{idid_weights.fixest}
\alias{idid_weights.fixest}
\title{Identifying Variation Weights for Fixed-effects OLS Fits}
\usage{
\method{idid_weights}{fixest}(reg, var_interest)
}
\arguments{
\item{reg}{A fitted \code{fixest::feols} model.}

\item{var_interest}{A string. The treatment or main variable of interest.}
}
\value{
A numeric vector representing the identifying variation weights.
}
\description{
\code{idid_weights} method for class "\code{fixest}".
}
\seealso{
Other idid_methods: 
\code{\link{idid_weights.lm}()}
}
\concept{idid_methods}
