% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iiv_weight.R
\name{idid_weights.lm}
\alias{idid_weights.lm}
\title{Identifying Variation Weights for Linear Model Fits}
\usage{
\method{idid_weights}{lm}(reg, var_interest)
}
\arguments{
\item{reg}{A fitted \code{lm} model.}

\item{var_interest}{A string. The treatment or main variable of interest.}
}
\value{
A numeric vector representing the identifying variation weights.
}
\description{
\code{idid_weights} method for class "\code{lm}".
}
\concept{idid_methods}
