% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_weights.R
\name{idid_weights}
\alias{idid_weights}
\title{Compute Identifying Variation Weights}
\usage{
idid_weights(reg, var_interest)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}
}
\value{
A numeric vector representing the identifying variation weights.
}
\description{
Compute regression weights describing the amount of variation used for
identification.
}
\details{
The weights correspond to the normalized leverage of each observation for
the variable of interest after partialling out all controls.

The function invokes particular methods which depend on the class of the
first argument.
}
\examples{
# example with a lm regression
reg_ex_lm <- ggplot2::txhousing |>
  lm(formula = volume ~ sales + listings + city + as.factor(date))

idid_weights(reg_ex_lm, "sales") |>
 head()

# example with a fixest regression
reg_ex_fixest <- ggplot2::txhousing  |>
  fixest::feols(fml = volume ~ sales + listings |  as.factor(date) + city)

idid_weights(reg_ex_fixest, "sales") |>
  head()

# example with a plm regression
reg_ex_plm <- ggplot2::txhousing  |>
  plm::plm(
    formula = volume ~ sales + listings,
    index = c("date", "city"),
    model = "within",
    effect = "twoways"
  )

idid_weights(reg_ex_plm, "sales") |>
  head()

}
