% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_weights_cat.R
\name{idid_weights_cat}
\alias{idid_weights_cat}
\title{Convert numeric vector of weights to categorical label}
\usage{
idid_weights_cat(weights)
}
\arguments{
\item{weights}{A numeric (positive) vector.}
}
\value{
A factor with levels "More than 50x smaller", "Between 10x and 50x smaller",
"Between 2x and 10x smaller", "Between 2x smaller and 2x larger",
"Between 2x and 10x larger", "Between 10x and 50x larger", and
"More than 50x larger", indicating the relative size of each element.
}
\description{
Categorize weights based on their size relative to the total weight.
}
