% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_weights_viz.R
\name{idid_weights_viz}
\alias{idid_weights_viz}
\title{Graph of of Identifying Variation Weights}
\usage{
idid_weights_viz(reg, var_interest, var_1, var_2, keep_labels = TRUE)
}
\arguments{
\item{data}{A regression object.}

\item{weights_vect}{A single string. Optional.}

\item{indiv_name}{One of the columns in \code{reg$model},
representing the individual dimension}

\item{time_name}{One of the columns in \code{reg$model},
representing the time dimension (if have a panel). Optional.}
}
\value{
An invisibly returned object using ggplot2. The function will error if
any of the provided variables are unavailable in \code{reg$model}.
}
\description{
Makes a heatmap to visualize the identifying variation weights
}
