% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_axes.R
\name{order_axes}
\alias{order_axes}
\title{Ordering axes}
\usage{
order_axes(df, name_var_x, name_var_y, order, by = "weight")
}
\arguments{
\item{df}{A dataframe to order}

\item{order}{A string (either "x", "y" or "xy") describing whether the
graph should be order, along the x or y axis or both.
If anything else is specified, no specific ordering will be applied.}
}
\description{
Orders axes depending on the order desired
Used in \code{idid_viz_weights} and \code{idid_viz_contrib}
}
\keyword{internal}
