% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sum_sf.R
\docType{data}
\name{StatSumSf}
\alias{StatSumSf}
\title{Automatic mean aggregation for sf}
\format{
An object of class \code{StatSumSf} (inherits from \code{StatSf}, \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatSumSf
}
\description{
A ggplot2 Stat that sums the values provided in z for each geometry
}
\keyword{internal}
