% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sum_tile.R
\docType{data}
\name{StatSumTile}
\alias{StatSumTile}
\title{Automatic aggregation and log-transform for tiles}
\format{
An object of class \code{StatSumTile} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatSumTile
}
\description{
A ggplot2 Stat that behaves like stat_summary_2d but automatically:
\enumerate{
\item Sums z-values per (x, y)
\item Applies the transformation log10(sum(z) * n_x * n_y)
}
}
\details{
summarise the weights by var_x\emph{vary_y groups (summing their values),
take the ratio of the weight of each group over
the average weight across groups (1/(n_x}n_y)) and then takes its log10
}
\keyword{internal}
