% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_idid.R
\name{theme_idid}
\alias{theme_idid}
\title{Create a custom theme for \code{ididvar} visualizations}
\usage{
theme_idid()
}
\value{
The specified \code{theme} object.
}
\description{
A function to return a custom ggplot2 theme.
}
