% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_contrib_threshold.R
\name{idid_contrib_threshold}
\alias{idid_contrib_threshold}
\title{Find a weight threshold below which observations "do not contribute"}
\usage{
idid_contrib_threshold(
  reg,
  var_interest,
  threshold_change = 0.05,
  search_step = 0.05
)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{threshold_change}{A double (between 0 and 1). The change threshold in
estimate and s.e. when dropping observations.}

\item{search_step}{A double (between 0 and 1). The additional proportion of
observations to drop in each iteration of the loop.}
}
\value{
A numeric. The largest weight below which one can drop observations
without altering the estimate by more than a proportion of
\code{threshold_change}.
}
\description{
Find a weight threshold below which removing observations does not change
the point estimate or the standard error of the estimate of interest by more
than a given proportion.
}
\details{
This function applies \code{ididvar::idid_drop_change} for a series of
proportions of observations removed, increasing this proportion by
\code{search_step} while the variation in the estimate and s.e. as compared
to the full sample is less than \code{threshold_change}.
}
\examples{
reg_ex <- ggplot2::economics |>
  lm(formula = unemploy ~ pce + uempmed + psavert + pop)

idid_contrib_threshold(reg_ex, "pce")
}
