% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_drop_change.R
\name{idid_drop_change}
\alias{idid_drop_change}
\title{Variation in estimate when dropping low weight observations}
\usage{
idid_drop_change(reg, var_interest, prop_drop)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{prop_drop}{Proportion of observations to drop}
}
\value{
A dataframe with one row and 3 columns:
\itemize{
\item \code{prop_drop}: the proportion of data dropped
\item \code{prop_change_est}: the relative change between the full sample and
droped estimates
\item \code{prop_change_se}: the relative change between the full sample and
droped s.e.
}
}
\description{
Computes a variation of the point estimate and standard errors of the
variable of interest when dropping low weight observations.
}
