% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_grouping_var.R
\name{idid_grouping_var}
\alias{idid_grouping_var}
\title{Compute the between-group variances in weights}
\usage{
idid_grouping_var(reg, var_interest, grouping_vars, ...)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{grouping_vars}{A string vector of variable names or \code{"everything"}.
The set of variables to group weights by. If \code{"everything"} passed,
group by each of the variables in the data set successively.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A dataframe with 2 columns:
\itemize{
\item \code{grouping_var} the names of grouping var
\item \code{between_var} the between-group variation in weights
}
}
\description{
Identify variables along which groups may contribute heterogeneously
to identification.
}
\details{
Identify variables for which grouping by this
variable yields the most heterogeneous between-groups differences in weights
(i.e. the largest between-groups variance).
}
\examples{
reg_ex <- ggplot2::txhousing |>
  lm(formula = log(sales) ~ median + listings + city + as.factor(date))

idid_grouping_var(reg_ex, "median", grouping_vars = c("city", "year", "month"))

#We can now then explore the weights along the dimension with the largest
#variance: year
idid_viz_weights(reg_ex, "median", year)
}
