% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_grouping_var.R
\name{idid_grouping_var}
\alias{idid_grouping_var}
\title{Computes the within-group and between-group weight variances}
\usage{
idid_grouping_var(reg, var_interest, grouping_vars = "everything")
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{grouping_vars}{A string vector of variable names or \code{"everything"}.
The set of variables to group weights by. If \code{"everything"} passed,
group by each of the variables in the data set successively.}
}
\value{
A dataframe with 3 columns:
\itemize{
\item \code{grouping_var} the names of grouping var
\item \code{within_var} the within-group variation in weights
\item \code{between_var} the between-group variation in weights
}
}
\description{
Groups weights by each of the variables listed in \code{grouping_vars},
and computes the within-group variances and between-group variances.
}
\details{
This function allows to identify variables for which grouping by this
variable yields the most homogeneous within group variance in weights or the
most heterogeneous between groups variance
}
\examples{
reg_ex <- ggplot2::economics |>
  lm(formula = unemploy ~ pce + uempmed + psavert + pop)

idid_grouping_var(reg_ex, "pce", grouping_vars = "everything")
}
