% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_partial_out.R
\name{idid_partial_out}
\alias{idid_partial_out}
\title{Partial out controls}
\usage{
idid_partial_out(reg, var_to_partial, ...)
}
\arguments{
\item{reg}{A regression object.}

\item{var_to_partial}{A string. The name of the variable to partial out.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A vector of the partialled out version of \code{var_to_partial}
}
\description{
Compute the partialled out version of a variable based on the regression
model provided.
}
\details{
The partialling out procedure is implemented using the same estimation
method as the one used in \code{reg}.
}
\examples{
# example with a lm regression
reg_ex_lm <- ggplot2::txhousing |>
  lm(formula = volume ~ sales + listings + city + as.factor(date))

idid_partial_out(reg_ex_lm, "sales") |>
 head()

idid_partial_out(reg_ex_lm, "volume") |>
 head()

# example with a fixest regression
reg_ex_fixest <- ggplot2::txhousing  |>
  fixest::feols(fml = volume ~ sales + listings |  as.factor(date) + city)

idid_partial_out(reg_ex_fixest, "sales") |>
  head()

idid_partial_out(reg_ex_fixest, "volume") |>
  head()
}
