% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_bivar.R
\name{idid_viz_bivar}
\alias{idid_viz_bivar}
\title{Visualization of the relationship between x and y, after partialling out}
\usage{
idid_viz_bivar(reg, var_interest, partial_iv = TRUE, ...)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{partial_iv}{A boolean. If TRUE, partials out the part of the
endogenous variables that are not explained by the instruments. If the
regression is not an IV (or not generated by \code{AER::ivreg} or \code{fixest::feols}
with an instrumented part), the value of \code{partial_iv} has no impact on the
outcome of the function.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A ggplot2 scatter plot and regression line of the relationship between the
outcome and the variable of interest, after partialling out controls.
}
\description{
Makes a bivariate graph to visualize the data and relationship between the
outcome and the variable of interest, after having partialLed out controls.
}
\examples{
reg_few_ctrl <- ggplot2::txhousing |>
  lm(formula = log(sales) ~ median)

idid_viz_bivar(reg_few_ctrl, "median")

reg_more_ctrl <- ggplot2::txhousing |>
  lm(formula = log(sales) ~ median + listings + city + as.factor(date))

idid_viz_bivar(reg_more_ctrl, "median")
}
