% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_contrib.R
\name{idid_viz_contrib}
\alias{idid_viz_contrib}
\title{Compute heatmap contribution weights}
\usage{
idid_viz_contrib(
  reg,
  var_interest,
  var_x,
  var_y,
  contrib_threshold,
  threshold_change = 0.05,
  keep_labels = TRUE,
  ...
)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The treatment or main variable of interest.}

\item{var_x}{A variable in the data set used in \code{reg}.}

\item{var_y}{A variable in the data set used in \code{reg}. If not specified,
produces a bar chart.}

\item{keep_labels}{A boolean (optional). If FALSE, remove y labels and ticks.}

\item{threshold_diff}{(Optional) A numeric value between 0 and 1 representing the maximum proportion of data points that can be removed before we suspect that our estimate is changing more than by a \code{threshold_diff} percentage.}
}
\value{
A ggplot object representing contribution heatmap. The function will print a warning message if \code{threshold_diff} is too low and not enough data would remain by dropping non-contributing cases.
}
\description{
Create a heatmap showing which observations contribute to the identification of certain variables.
}
