% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_cumul.R
\name{idid_viz_cumul}
\alias{idid_viz_cumul}
\title{Visualization of the cumulative distribution of weights}
\usage{
idid_viz_cumul(reg, var_interest, prop_weights = 0.2, color = "#300D49", ...)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{prop_weights}{A numeric (between 0 and 1). A proportion of weights
that to consider as contributing the most to identification.
Used to build the subtitle.}

\item{color}{A string. Color of the graph.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A ggplot2 graph of the cumulative distribution of weights.
}
\description{
Visualization of the cumulative distribution of weights
}
\examples{
reg_ex <- ggplot2::txhousing |>
  lm(formula = volume ~ sales + listings + city + as.factor(date))

idid_weights(reg_ex, "sales") |>
  idid_viz_cumul()
}
