% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_cumul.R
\name{idid_viz_cumul}
\alias{idid_viz_cumul}
\title{Visualization of the cumulative distribution of weights}
\usage{
idid_viz_cumul(weights, prop_weights = 0.25)
}
\arguments{
\item{weights}{A numeric vector (e.g., weights or probabilities).}

\item{prop_weights}{Optional. One of \code{0}, \code{1}, or values in-between. The contribution threshold above which to display contributions ("tail") individually.}
}
\value{
An idiD Lörentz plot
}
\description{
Visualization of the cumulative distribution of weights
}
