% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_drop_change.R
\name{idid_viz_drop_change}
\alias{idid_viz_drop_change}
\title{Visualize changes in point estimate and s.e. when dropping observations}
\usage{
idid_viz_drop_change(
  reg,
  var_interest,
  threshold_change = 0.05,
  search_step = 0.05,
  search_start = search_step,
  search_end = 1 - search_step,
  color = "#300D49",
  ...
)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{threshold_change}{A double (between 0 and 1). The change threshold in
estimate and s.e. when dropping observations.}

\item{search_step}{A double (between 0 and 1). The additional proportion of
observations to drop in each iteration of the loop.}

\item{search_start}{A numeric (between 0 and 1). Proportion of observation to
drop in the first step of the loop in \code{idid_contrib_threshold}.}

\item{search_end}{A numeric (between 0 and 1). Proportion of observation to
drop in the last step of the loop in \code{idid_contrib_threshold}.}

\item{color}{A string. Color of the graph.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A ggplot2 object.

The name of the facet variable is \code{measure}.
}
\description{
Make a faceted line graph describing the percentage variation of the point
estimate and s.e. of interest as a function of the proportion of
observations dropped.
}
\examples{
reg_ex <- ggplot2::txhousing |>
  lm(formula = log(sales) ~ median + listings + city + as.factor(date))

idid_viz_drop_change(reg_ex, "median", search_end = 0.6)
}
