% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_viz_weights.R
\name{idid_viz_weights}
\alias{idid_viz_weights}
\title{Visualization of Identifying Variation Weights}
\usage{
idid_viz_weights(
  reg,
  var_interest,
  var_x,
  var_y,
  colors = c("#C25807", "#FBE2C5", "#300D49"),
  keep_labels = TRUE,
  ...
)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{var_x}{A variable in the data set used in \code{reg} to plot on the
x-axis.}

\item{var_y}{A variable in the data set used in \code{reg} to plot on the
y-axis (optional). If not specified, produces a bar chart.}

\item{colors}{A string vector of colors for the palette.
I recommend to pass a vector of 3 distinct colors, with a lighter color in
the middle, constituting a diverging scale. It allows a clear distinction
between contributing and non contributing observations.}

\item{keep_labels}{A boolean (optional). If FALSE, remove y labels and ticks.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A ggplot2 graph of identifying variation weights.

If var_y is specified, returns a heatmap whose color represents a
categorized version of the identifying variation weights (the categorization
prevents the vizualisation from being driven by extremely large weights).

If var_y is not specified, returns a bar chart representing the weights in
each group.
}
\description{
Makes a graph to visualize the identifying variation weights (a heatmap or
a bar chart, depending on the number of dimensions specified)
}
\details{
If there is more than on observation by group (i.e. for one category in
var_x or in var_x x var_y), individual weights are added for that group,
removing missing values.
}
\examples{
# example with a lm regression and one dimension
reg_ex_lm <- ggplot2::economics |>
  transform(year = substr(date, 1, 4) |> as.numeric()) |>
  lm(formula = unemploy ~ pce + uempmed + psavert + pop + year)

idid_viz_weights(reg_ex_lm, "pce", var_x = year) +
  ggplot2::labs(x = NULL)

# example with a fixest regression and two dimensions
reg_ex_fixest <- ggplot2::txhousing |>
  fixest::feols(fml = volume ~ sales + listings |  as.factor(date) + city)

idid_viz_weights(reg_ex_fixest, "sales", date, city) +
  ggplot2::labs(x = NULL, y = NULL)
}
