% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_weights.R
\name{idid_weights}
\alias{idid_weights}
\title{Identifying Variation Weights}
\usage{
idid_weights(reg, var_interest)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The treatment or main variable of interest.}
}
\value{
A numeric vector representing the identifying variation weights.
}
\description{
A generic function used to compute regression weights
describing the amount of variation that is used for the identification of
a treatment effect.
}
\details{
The weights correspond to the normalized leverage of each observation for
the variable of interest after partialling out all controls.

The function invokes particular methods which depend on the class of the
first argument.
}
