% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_weights.R
\name{idid_weights}
\alias{idid_weights}
\title{Compute Identifying Variation Weights}
\usage{
idid_weights(reg, var_interest, partial_iv = TRUE, ...)
}
\arguments{
\item{reg}{A regression object.}

\item{var_interest}{A string. The name of the main variable of interest.}

\item{partial_iv}{A boolean. If TRUE, partials out the part of the
endogenous variables that are not explained by the instruments. If the
regression is not an IV (or not generated by \code{AER::ivreg} or \code{fixest::feols}
with an instrumented part), the value of \code{partial_iv} has no impact on the
outcome of the function.}

\item{...}{Additional elements to pass to the regression function when
partialling out controls.}
}
\value{
A numeric vector representing the identifying variation weights.
}
\description{
Compute regression weights describing the amount of variation used for
identification.
}
\details{
The weights correspond to the normalized leverage of each observation for
the variable of interest after partialling out all controls.

They are computed by re-running the regression provided but replacing the
independent variable by the variable of interest (and removing it from the
set of regressors).

If the nature of the independent variable and of the variable of interest
are different, one may want to change the estimation.
For instance, if the independent variable is discrete and estimated via glm
with the argument \code{family = "binomial"} and the dependent variable is
continuous, one may want to provide the argument \code{family = "gaussian"}
to \code{idid_weights}.
}
\examples{
reg_ex_lm <- ggplot2::txhousing |>
  lm(formula = log(sales) ~ median + listings + city + as.factor(date))

idid_weights(reg_ex_lm, "median") |>
 head()

}
