% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ididvar-package.R
\docType{package}
\name{ididvar-package}
\alias{ididvar}
\alias{ididvar-package}
\title{ididvar: Identifying the Identifying Variation in a Regression}
\description{
Provides a set of tools to identify where the identifying variation in a regression comes from.
}
\author{
\strong{Maintainer}: Vincent Bagilet \email{vincent.bagilet@ens-lyon.fr}

}
\keyword{internal}
