% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sum_tile.R
\docType{data}
\name{StatSumTile}
\alias{StatSumTile}
\title{Automatic mean aggregation for tiles}
\format{
An object of class \code{StatSumTile} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatSumTile
}
\description{
A ggplot2 Stat that sums the values provided in z for each tile
}
\keyword{internal}
