% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_idid.R
\name{theme_idid}
\alias{theme_idid}
\title{Create a theme with minimal background and a customized legend}
\usage{
theme_idid()
}
\arguments{
\item{None}{}
}
\value{
The specified \code{theme} object. This will error invisibly if used incorrectly.
}
\description{
A function to return a custom ggplot2 theme.
}
