% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_idid.R
\name{theme_idid}
\alias{theme_idid}
\title{Create a custom theme for \code{ididvar} visualizations}
\usage{
theme_idid(...)
}
\arguments{
\item{...}{Additional arguments to be passed to the \code{theme} function}
}
\value{
The specified \code{theme} object.
}
\description{
A function to return a custom ggplot2 theme.
}
\examples{
ggplot2::mpg |>
  ggplot2::ggplot(ggplot2::aes(x = displ, hwy, color = cyl)) +
  ggplot2::geom_point() +
  ggplot2::labs(
    title = "A wonderful title",
    subtitle = "Something to add"
  ) +
  theme_idid()
}
